Revert "RESTful end of form navigation"
