+++ src/main/java/application/FormController.java
@@ -25,7 +25,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import repo.SerializableMenuSession;
 import services.CategoryTimingHelper;
-import services.MenuSessionRunnerService;
 import services.SubmitService;
 import services.XFormService;
 import session.FormSession;
@@ -55,6 +54,7 @@ public class FormController extends AbstractBaseController{
 
     @Autowired
     private CategoryTimingHelper categoryTimingHelper;
+
 
     @Value("${commcarehq.host}")
     private String host;
@@ -181,8 +181,10 @@ public class FormController extends AbstractBaseController{
 
     private Object doEndOfFormNav(SerializableMenuSession serializedSession) throws Exception {
         log.info("End of form navigation with serialized menu session: " + serializedSession);
-        MenuSession menuSession = menuSessionFactory.buildSession(serializedSession);
-        return runnerService.resolveFormGetNext(menuSession);
+        MenuSession menuSession = new MenuSession(serializedSession, installService, restoreFactory, host);
+        Object nextScreen = resolveFormGetNext(menuSession);
+        menuSessionRepo.save(new SerializableMenuSession(menuSession));
+        return nextScreen;
     }
 
     /**
