+++ src/main/java/application/WebAppContext.java
@@ -335,10 +335,4 @@ public class WebAppContext extends WebMvcConfigurerAdapter {
     public FormplayerFormSendCalloutHandler formSendCalloutHandler() {
         return new FormplayerFormSendCalloutHandler();
     }
-
-    @Bean
-    public MenuSessionRunnerService menuSessionRunnerService() {return new MenuSessionRunnerService();}
-
-    @Bean
-    public MenuSessionFactory menuSessionFactory() {return new MenuSessionFactory();}
 }
