+++ src/main/java/beans/InstallFromSessionRequestBean.java
@@ -1,5 +1,8 @@
 package beans;
 
+import com.fasterxml.jackson.annotation.JsonGetter;
+import com.fasterxml.jackson.annotation.JsonSetter;
+
 /**
  * Created by benrudolph on 2/4/16.
  *
