+++ src/main/java/beans/menus/CommandListResponseBean.java
@@ -35,6 +35,8 @@ public class CommandListResponseBean extends MenuBean {
             commands[i] = command;
         }
         this.setCommands(commands);
+        this.setMenuSessionId(menuSessionId);
+
         String menuId = "root";
         if (session.getCommand() != null) {
             menuId = session.getCommand();
