+++ src/main/java/beans/menus/MenuBean.java
@@ -11,6 +11,7 @@ public class MenuBean extends BaseResponseBean {
 
     private String[] locales;
     private String[] breadcrumbs;
+    private String menuSessionId;
     private EntityDetailResponse persistentCaseTile;
 
     @Override
@@ -33,6 +34,13 @@ public class MenuBean extends BaseResponseBean {
 
     public void setBreadcrumbs(String[] breadcrumbs) {
         this.breadcrumbs = breadcrumbs;
+    }
+    public String getMenuSessionId() {
+        return menuSessionId;
+    }
+
+    public void setMenuSessionId(String menuSessionId) {
+        this.menuSessionId = menuSessionId;
     }
 
     public EntityDetailResponse getPersistentCaseTile() {
