+++ src/test/java/tests/Enikshay2bTests.java
@@ -48,22 +48,15 @@ public class Enikshay2bTests extends BaseTestClass {
         LinkedHashMap commandsRaw = (LinkedHashMap) submitResponse.getNextScreen();
         String jsonString = new JSONObject(commandsRaw).toString();
         CommandListResponseBean commandResponse = mapper.readValue(jsonString, CommandListResponseBean.class);
-        String[] selections = commandResponse.getSelections();
-        assert  selections.length == 2;
-        String commandSelection = selections[0];
-        String caseSelection = selections[1];
-        assert "0".equals(commandSelection);
-        String[] newSelections = new String[] {commandSelection, caseSelection, "10"};
-        EntityListResponse sortedEntityResponse = sessionNavigate(newSelections,
-                "enikshay-2b",
+        EntityListResponse sortedEntityResponse = sessionNavigateWithId(new String[] {"10"},
+                commandResponse.getMenuSessionId(),
                 4,
                 EntityListResponse.class);
-        EntityListResponse entityResponse = sessionNavigate(newSelections,
-                "enikshay-2b",
+        EntityListResponse entityResponse = sessionNavigateWithId(new String[] {"10"},
+                commandResponse.getMenuSessionId(),
                 EntityListResponse.class);
-        newSelections = new String[] {commandSelection, caseSelection, "10", "4edc7248-bfef-4f47-95e0-6b1704981a3f"};
-        NewFormResponse newFormResponse2 = sessionNavigate(newSelections,
-                "enikshay-2b",
+        NewFormResponse newFormResponse2 = sessionNavigateWithId(new String[] {"10", "4edc7248-bfef-4f47-95e0-6b1704981a3f"},
+                commandResponse.getMenuSessionId(),
                 NewFormResponse.class);
     }
