--- src/test/resources/requests/evaluate_xpath/evaluate_xpath_menu.json
@@ -1,9 +1,0 @@
-{
-  "username": "loaduser",
-  "password": "loadpass",
-  "domain": "loaddomain",
-  "app_id": "loadappid",
-  "selections": [2, 1],
-  "installReference": "archives/basic.ccz",
-  "xpath": "count(instance('casedb')/casedb/case)"
-}
