--- src/test/resources/requests/submit/submit_enikshay_private_0.json
@@ -1,17 +1,0 @@
-{
-  "action": "submit-all",
-  "answers": {
-    "0,0": "1",
-    "0,1": "2017-12-04",
-    "0,3": "[1]",
-    "1,0,0": "NIKITA",
-    "1,0,1": "VERMA",
-    "1,0,2": "9076466580",
-    "1,1,0": "21",
-    "1,3": "[1]",
-    "1,4,0":"2"
-  },
-  "domain": "enikshay_privatedomain",
-  "username": "enikshay_privateusername",
-  "prevalidated": true
-}
