--- src/test/resources/requests/submit/submit_enikshay_private_1.json
@@ -1,14 +1,0 @@
-{
-  "action": "submit-all",
-  "answers": {
-    "0,0": "2017-12-04",
-    "0,1": "13",
-    "0,2": "11",
-    "0,4": "1",
-    "1,0": "1",
-    "2,0": "1"
-  },
-  "domain": "enikshay_privatedomain",
-  "username": "enikshay_privateusername",
-  "prevalidated": true
-}
